requirements-agent.txt
# PDF2BPMN Agent Server 의존성
# pip install -r requirements-agent.txt

# ProcessGPT SDK (에이전트 프레임워크)
# PyPI: https://pypi.org/project/process-gpt-agent-sdk/
process-gpt-agent-sdk==0.4.13

# Supabase (데이터베이스 연동)
supabase>=2.0.0

# HTTP 클라이언트
httpx>=0.25.0

# 환경변수 로드
python-dotenv>=1.0.0

# 기타 유틸리티
pydantic>=2.0.0
